Moving common enums into common.h.
