+++ include/network_interface/network_utils.h
@@ -8,22 +8,19 @@
 #ifndef NETWORK_INTERFACE_NETWORK_UTILS_H
 #define NETWORK_INTERFACE_NETWORK_UTILS_H
 
-#include <cstddef>
+#include "network_interface/common.h"
+
 #include <cstdint>
 #include <vector>
 #include <typeinfo>
-#include <cstring>
+#include <string>
+#include <algorithm>
 #include <type_traits>
 
 namespace AS
 {
 namespace Network
 {
-enum ByteOrder
-{
-  BE = 0,
-  LE
-};
 
 inline bool system_is_big_endian()
 {
@@ -56,9 +53,9 @@ T read_le(const std::vector<uint8_t>& bufArray,
   T retVal = 0;
 
   if (system_is_big_endian())
-    std::memcpy(&retVal, &rcvData + sizeof(uint64_t) - sizeof(T), sizeof(T));
+    std::copy(&retVal, &rcvData + sizeof(uint64_t) - sizeof(T), sizeof(T));
   else
-    std::memcpy(&retVal, &rcvData, sizeof(T));
+    std::copy(&retVal, &rcvData, sizeof(T));
 
   retVal *= static_cast<T>(factor);
   retVal += valueOffset;
@@ -113,9 +110,9 @@ T read_be(const std::vector<uint8_t>& bufArray,
   T retVal;
 
   if (system_is_big_endian())
-    std::memcpy(&retVal, &rcvData + sizeof(uint64_t) - sizeof(T), sizeof(T));
+    std::copy(&retVal, &rcvData + sizeof(uint64_t) - sizeof(T), sizeof(T));
   else
-    std::memcpy(&retVal, &rcvData, sizeof(T));
+    std::copy(&retVal, &rcvData, sizeof(T));
 
   retVal *= (T) factor;
   retVal += valueOffset;
@@ -181,6 +178,29 @@ inline int32_t find_magic_word(const std::vector<uint8_t>& in, const size_t& mag
   // Just in case
   return -1;
 };
+
+std::string return_status_desc(const ReturnStatuses &ret)
+{
+  if (ret == ReturnStatuses::INIT_FAILED)
+    return "Initialization of the network interface failed.";
+  else if (ret == ReturnStatuses::BAD_PARAM)
+    return "A bad parameter was provided to the network interface during initalization.";
+  else if (ret == ReturnStatuses::SOCKET_ERROR)
+    return "A socket error was encountered.";
+  else if (ret == ReturnStatuses::SOCKET_CLOSED)
+    return "Socket is not currently open.";
+  else if (ret == ReturnStatuses::NO_MESSAGES_RECEIVED)
+    return "No messages were received on the interface.";
+  else if (ret == ReturnStatuses::READ_FAILED)
+    return "A read operation failed on the network interface.";
+  else if (ret == ReturnStatuses::WRITE_FAILED)
+    return "A write operation failed on the network interface.";
+  else if (ret == ReturnStatuses::CLOSE_FAILED)
+    return "Closing the network failed.";
+  else
+    return "Undefined error.";
+}
+
 }  // namespace Network
 }  // namespace AS
